requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
docx2txt